pyproject.toml
[project]
name = "commuse"
version = "0.1.0"
description = "AI-powered semantic commit message generator"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]
requires-python = ">=3.8"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[project.scripts]
commuse = "commuse.cli:app"